+++ src/faker_to_snowflake/config.py
@@ -1,4 +1,3 @@
 secret_name = 'gmdata_snowflake'
 region = 'us-east-1'
-profile_name = 'snowpipe-tst'
